Alter pubic interface for signal context.
Change ContextWithCancelSignals and ProgramContext to have an API that
looks more familiar to the standard library context.WithCancel and
context.WithDeadline. This also allows SignalContext to be made private.

Move ExitHint to be a function, not a SignalContext method. Also
introduce an ExitHinter interface that can be implemented by error types
to suggest appropriate error codes when there is no signal received.

Update example code.